SASS-5371: Relocated files to correct folders, updated routes and tests
