+++ app/viewmodels/checkAnswers/UkDividendsSharesAndLoansFromLimitedCompaniesSummary.scala
@@ -16,9 +16,8 @@
 
 package viewmodels.checkAnswers
 
-import controllers.routes
 import models.{CheckMode, UserAnswers}
-import pages.UkDividendsSharesAndLoansFromLimitedCompaniesPage
+import pages.propertypensionsinvestments.UkDividendsSharesAndLoansFromLimitedCompaniesPage
 import play.api.i18n.Messages
 import play.twirl.api.HtmlFormat
 import uk.gov.hmrc.govukfrontend.views.viewmodels.content.HtmlContent
@@ -45,8 +44,11 @@ object UkDividendsSharesAndLoansFromLimitedCompaniesSummary  {
           key     = "ukDividendsSharesAndLoansFromLimitedCompanies.checkYourAnswersLabel",
           value   = value,
           actions = Seq(
-            ActionItemViewModel("site.change", routes.UkDividendsSharesAndLoansFromLimitedCompaniesController.onPageLoad(CheckMode, answers.taxYear).url)
-              .withVisuallyHiddenText(messages("ukDividendsSharesAndLoansFromLimitedCompanies.change.hidden"))
+            ActionItemViewModel(
+              "site.change",
+              controllers.propertypensionsinvestments.routes.UkDividendsSharesAndLoansFromLimitedCompaniesController
+                .onPageLoad(CheckMode, answers.taxYear).url)
+                .withVisuallyHiddenText(messages("ukDividendsSharesAndLoansFromLimitedCompanies.change.hidden"))
           )
         )
     }
