+++ conf/app.routes
@@ -155,7 +155,7 @@ POST       /:taxYear/property-pensions-investments/uk-interest                  
 GET        /:taxYear/property-pensions-investments/change-uk-interest                                                   controllers.propertypensionsinvestments.UkInterestController.onPageLoad(mode: Mode = CheckMode, taxYear: Int)
 POST       /:taxYear/property-pensions-investments/change-uk-interest                                                   controllers.propertypensionsinvestments.UkInterestController.onSubmit(mode: Mode = CheckMode, taxYear: Int)
 
-GET        /:taxYear/property-pensions-investments/uk-dividends-shares-loans          controllers.UkDividendsSharesAndLoansFromLimitedCompaniesController.onPageLoad(mode: Mode = NormalMode, taxYear: Int)
-POST       /:taxYear/property-pensions-investments/uk-dividends-shares-loans          controllers.UkDividendsSharesAndLoansFromLimitedCompaniesController.onSubmit(mode: Mode = NormalMode, taxYear: Int)
-GET        /:taxYear/property-pensions-investments/change-uk-dividends-shares-loans   controllers.UkDividendsSharesAndLoansFromLimitedCompaniesController.onPageLoad(mode: Mode = CheckMode, taxYear: Int)
-POST       /:taxYear/property-pensions-investments/change-uk-dividends-shares-loans   controllers.UkDividendsSharesAndLoansFromLimitedCompaniesController.onSubmit(mode: Mode = CheckMode, taxYear: Int)
+GET        /:taxYear/property-pensions-investments/uk-dividends-shares-loans          controllers.propertypensionsinvestments.UkDividendsSharesAndLoansFromLimitedCompaniesController.onPageLoad(mode: Mode = NormalMode, taxYear: Int)
+POST       /:taxYear/property-pensions-investments/uk-dividends-shares-loans          controllers.propertypensionsinvestments.UkDividendsSharesAndLoansFromLimitedCompaniesController.onSubmit(mode: Mode = NormalMode, taxYear: Int)
+GET        /:taxYear/property-pensions-investments/change-uk-dividends-shares-loans   controllers.propertypensionsinvestments.UkDividendsSharesAndLoansFromLimitedCompaniesController.onPageLoad(mode: Mode = CheckMode, taxYear: Int)
+POST       /:taxYear/property-pensions-investments/change-uk-dividends-shares-loans   controllers.propertypensionsinvestments.UkDividendsSharesAndLoansFromLimitedCompaniesController.onSubmit(mode: Mode = CheckMode, taxYear: Int)
